+++ lib/src/main/java/ai/freeplay/client/resources/datasets/Pagination.java
@@ -1,0 +1,38 @@
+package ai.freeplay.client.resources.datasets;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class Pagination {
+    private int page;
+    private int pageSize;
+    private boolean hasNext;
+
+    public Pagination() {
+    }
+
+    public void setPage(int page) {
+        this.page = page;
+    }
+
+    public void setPageSize(int pageSize) {
+        this.pageSize = pageSize;
+    }
+
+    public void setHasNext(boolean hasNext) {
+        this.hasNext = hasNext;
+    }
+
+    public int getPage() {
+        return page;
+    }
+
+    public int getPageSize() {
+        return pageSize;
+    }
+
+    public boolean isHasNext() {
+        return hasNext;
+    }
+}